With ptrhead, Thread::currentTid() returns a pointer to a struct which leads to a compile error when assigning it to an long. Thus _ostid type should be changed to IntPtr and the assignment should use a reinterpret_cast.

+++ Foundation/include/Poco/Message.h
@@ -199,7 +199,7 @@ private:
 	Priority    _prio;
 	Timestamp   _time;
 	int         _tid;
-	int         _ostid;
+	IntPtr      _ostid;
 	std::string _thread;
 	long        _pid;
 	const char* _file;
@@ -247,7 +247,7 @@ inline long Message::getTid() const
 }
 
 
-inline long Message::getOsTid() const
+inline IntPtr Message::getOsTid() const
 {
 	return _ostid;
 }
